- fix compiler warning
git-svn-id: http://streamboard.de.vu/svn/oscam/trunk@8623 4b0bc96b-bc66-0410-9d44-ebda105a78c1

+++ module-dvbapi.c
@@ -1218,7 +1218,7 @@ void dvbapi_read_priority(void) {
 		char c_srvid[34];
 		c_srvid[0]='\0';
 		uint32_t caid=0, provid=0, srvid=0, ecmpid=0;
-		int32_t chid=-1; //chid=0 is a valid chid
+		int16_t chid=-1; //chid=0 is a valid chid
 		ret = sscanf(str1, "%4x:%6x:%33[^:]:%4x:%4x", &caid, &provid, c_srvid, &ecmpid, &chid);
 		if (ret < 1) {
 			cs_debug_mask(D_DVBAPI, "Error in oscam.dvbapi: ret=%d | %c: %04X %06X %s %04X %04X",
